+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/seo-assistant/use-keywords-step.tsx
@@ -1,4 +1,10 @@
-import { createInterpolateElement, useCallback, useEffect, useState } from '@wordpress/element';
+import {
+	createInterpolateElement,
+	useCallback,
+	useEffect,
+	useRef,
+	useState,
+} from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { useMessages } from './wizard-messages';
 import type { Step } from './types';
@@ -82,5 +88,6 @@ export const useKeywordsStep = (): Step => {
 		value,
 		setValue,
 		onStart,
+		inputRef: useRef( null ),
 	};
 };
